+++ bin/templates/components.tsx
@@ -1,7 +1,10 @@
+/**
+ * THIS FILE IS GENERATED
+ */
 ${components.map(component => "import " + component.capitalized + " from '" + component.root + component.name + "'").join('\n')}
 
 import branding from '_utils/branding'
-import { flush, flushToHTML } from 'styled-jsx/server'
+import flush, { flushToHTML } from 'styled-jsx/server'
 
 export * from 'icon/index'