+++ src/icon/index.tsx
@@ -1,31 +1,43 @@
+/**
+ * THIS FILE IS GENERATED
+ */
 import ArrowIcon from './arrowIcon'
+import BankIcon from './bankIcon'
+import BellIcon from './bellIcon'
+import BubbleIcon from './bubbleIcon'
 import CheckIcon from './checkIcon'
 import ChevronIcon from './chevronIcon'
 import CircleIcon from './circleIcon'
 import ClockIcon from './clockIcon'
+import CoinIcon from './coinIcon'
 import ComfortIcon from './comfortIcon'
+import CrossDiscIcon from './crossDiscIcon'
 import CrossIcon from './crossIcon'
 import CrosshairIcon from './crosshairIcon'
-import DecreaseIcon from './decreaseIcon'
 import DoubleArrowIcon from './doubleArrowIcon'
 import EyeIcon from './eyeIcon'
 import FacebookIcon from './facebookIcon'
-import IncreaseIcon from './increaseIcon'
+import HomeIcon from './homeIcon'
 import InfoIcon from './infoIcon'
 import InstagramIcon from './instagramIcon'
 import LadyIcon from './ladyIcon'
 import LightningIcon from './lightningIcon'
+import MinusIcon from './minusIcon'
 import MusicIcon from './musicIcon'
+import NewspaperIcon from './newspaperIcon'
 import OdnoklassnikiIcon from './odnoklassnikiIcon'
 import PetIcon from './petIcon'
 import PinIcon from './pinIcon'
 import PinterestIcon from './pinterestIcon'
+import PlusIcon from './plusIcon'
+import ProfileIcon from './profileIcon'
 import ProximityIcon from './proximityIcon'
 import QuestionIcon from './questionIcon'
 import SearchIcon from './searchIcon'
 import SmokeIcon from './smokeIcon'
 import SpeakIcon from './speakIcon'
 import StarIcon from './starIcon'
+import TicketIcon from './ticketIcon'
 import TrashIcon from './trashIcon'
 import TwitterIcon from './twitterIcon'
 import VkontakteIcon from './vkontakteIcon'
@@ -33,33 +45,42 @@ import YoutubeIcon from './youtubeIcon'
 
 export {
   ArrowIcon,
+  BankIcon,
+  BellIcon,
+  BubbleIcon,
   CheckIcon,
   ChevronIcon,
   CircleIcon,
   ClockIcon,
+  CoinIcon,
   ComfortIcon,
+  CrossDiscIcon,
   CrossIcon,
   CrosshairIcon,
-  DecreaseIcon,
   DoubleArrowIcon,
   EyeIcon,
   FacebookIcon,
-  IncreaseIcon,
+  HomeIcon,
   InfoIcon,
   InstagramIcon,
   LadyIcon,
   LightningIcon,
+  MinusIcon,
   MusicIcon,
+  NewspaperIcon,
   OdnoklassnikiIcon,
   PetIcon,
   PinIcon,
   PinterestIcon,
+  PlusIcon,
+  ProfileIcon,
   ProximityIcon,
   QuestionIcon,
   SearchIcon,
   SmokeIcon,
   SpeakIcon,
   StarIcon,
+  TicketIcon,
   TrashIcon,
   TwitterIcon,
   VkontakteIcon,