+++ src/text/story.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 
 import { storiesOf } from '@storybook/react'
 import { withInfo } from '@storybook/addon-info'
-import { selectV2, text, withKnobs } from '@storybook/addon-knobs'
+import { selectV2, text, withKnobs, boolean } from '@storybook/addon-knobs'
 
 import Text, { TextTagType, TextDisplayType } from 'text'
 
@@ -15,6 +15,7 @@ stories.add(
     <Text
       display={selectV2('Display', TextDisplayType, TextDisplayType.BODY)}
       tag={selectV2('Tag', TextTagType, TextTagType.SPAN)}
+      newlineToBr={boolean('newlineToBr', true)}
     >
       {text('Text', 'This is an example')}
     </Text>